backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
typing-extensions==4.8.0
langchain
sentence-transformers
chromadb
faiss-cpu
openai
pymupdf
pdfminer.six
pytesseract
Pillow
easyocr
python-dotenv
googletrans==4.0.0rc1
pymongo
indic-transliteration
tqdm
requests
transformers
